Move `api.go` to is own package and handle ListenAndServe error
This commit moves `api.go` to its own package so mock-es can be
imported by other Go programs.

The error returned by `http.ListenAndServe` is now handled and leads
to a log fatal.
